Level Grind: Fix 'lead can battle' check to ignore status conditions and health
Fixes #648
